requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
boto3
pydantic-settings